Automatically assign PBR textures from base texture name in BaseMaterial3D
This works based on a common set of texture name conventions that most
texture sites out there follow. It works with both StandardMaterial3D
and ORMMaterial3D. If mismatched textures are detected (compared
to the current material type), a notice is displayed.

Since this is implemented in EditorInspectorPluginMaterial,
this only affects textures set in the inspector, similar to the existing
behavior where `roughness` and `metallic` are automatically set to `1.0`
when assigning a texture to them.